refactor(source-generators): extract generator helpers and add AtcXamlToolkit0010
  - Decompose `ViewModelGenerator`, `FrameworkElementGenerator`, and
    `ObservableDtoViewModelGenerator` into slim Initialize/Execute shells,
    moving predicate/transform/diagnostic logic into dedicated helpers:
    `ViewModelGeneratorHelper`, `ViewModelDiagnosticHelper`,
    `FrameworkElementGeneratorHelper`, `ObservableDtoViewModelGeneratorHelper`
  - Promote `INotifyPropertyChangedTarget` into its own model file
  - Add diagnostic `AtcXamlToolkit0010` (Warning) emitted when `[RoutedEvent]`
    is detected on a WinUI 3 or Avalonia compilation, replacing the previous
    silent skip with an actionable build-time message
  - Wire the new diagnostic via `RegisterRoutedEventOnNonWpfDiagnostic` in
    `FrameworkElementGeneratorHelper`, backed by new `RoutedEventLocations`
    pipeline model and `DiagnosticFactory.CreateRoutedEventOnNonWpf`
  - Update `docs/SourceGenerators/Readme.md` and `RoutedEvents.md` to document
    the new diagnostic, with example warning text
  - Add `FrameworkElementGeneratorTests_RoutedEvent_NonWpfDiagnostic` covering
    WinUI 3 and Avalonia emission paths
