+++ docs/SourceGenerators/Readme.md
@@ -74,6 +74,7 @@ The generator surfaces these diagnostics at compile time so common misuses are c
 | `AtcXamlToolkit0007` | Warning | `[ComputedProperty]` is on a property whose getter doesn't reference any other property on the class — the generator silently filters such cases out, so the attribute does nothing. | Reference at least one `[ObservableProperty]` field or another property in the getter, or remove `[ComputedProperty]`. |
 | `AtcXamlToolkit0008` | Warning | Two or more `[ComputedProperty]` getters mutually reference each other (`A => B + 1` and `B => A + 1`, or any longer cycle). The generator's incremental invalidation graph doesn't trigger an update loop, but evaluating any of them at runtime infinitely recurses until the stack overflows. The diagnostic message includes the shortest cycle path. | Break the cycle by extracting shared computation into a non-computed helper, or by making one of the properties an `[ObservableProperty]`-backed value. |
 | `AtcXamlToolkit0009` | Warning | `[NotifyDataErrorInfo]` is on a field whose containing class doesn't transitively derive from `ObservableValidator` (the only base type that exposes the protected `ValidateProperty` method the generator calls in the setter). Combinations like `[INotifyPropertyChanged]` (plain class) + `[NotifyDataErrorInfo]` would otherwise produce an opaque `CS0103: name 'ValidateProperty' does not exist` inside the generated file. | Change the base class to `ObservableValidator` or `ViewModelBase`, or remove `[NotifyDataErrorInfo]`. |
+| `AtcXamlToolkit0010` | Warning | `[RoutedEvent]` is on a field in a project that targets WinUI 3 or Avalonia. WPF's `RoutedEvent` / `EventManager` infrastructure does not exist on those platforms, so the generator skips routed-event emission — the field would otherwise look like it should produce an event but produce nothing. | Replace the field with a standard CLR event (e.g. `public event RoutedEventHandler ItemSelected;`) or remove the attribute. |
 
 ## Architecture & decisions
