+++ src/Atc.XamlToolkit.SourceGenerators/GlobalUsings.cs
@@ -11,6 +11,7 @@ global using Atc.XamlToolkit.SourceGenerators.Extensions;
 global using Atc.XamlToolkit.SourceGenerators.Extensions.Builder;
 global using Atc.XamlToolkit.SourceGenerators.Extensions.CodeAnalysis;
 global using Atc.XamlToolkit.SourceGenerators.Factories;
+global using Atc.XamlToolkit.SourceGenerators.Generators.Helpers;
 global using Atc.XamlToolkit.SourceGenerators.Inspectors;
 global using Atc.XamlToolkit.SourceGenerators.Inspectors.Helpers;
 global using Atc.XamlToolkit.SourceGenerators.Models;
