+++ src/Atc.XamlToolkit.SourceGenerators/Models/FrameworkElement/RoutedEventLocations.cs
@@ -1,0 +1,4 @@
+namespace Atc.XamlToolkit.SourceGenerators.Models.FrameworkElement;
+
+internal sealed record RoutedEventLocations(
+    EquatableArray<(string FieldName, Location Location)> Locations);
